+++ packages/memory-store/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @melandlabs/memory-store
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @melandlabs/env-config@0.4.0
+  - @melandlabs/ai@0.3.1
+
 ## 0.3.0
 
 ### Minor Changes